Split the useUserInfoChange to separate hooks useCardChange and useShippingAddressChange
